던전 1, 2 추가
- 발판, 디스펜서 등 던전 기믹 스크립트 추가
- 투사체 로직 변경

+++ Assets/Scripts/Enemy/EnemyTurretController.cs
@@ -18,7 +18,7 @@ public class EnemyTurretController : MonoBehaviourPun
     [SerializeField] private Transform muzzlePoint;
     [SerializeField] private float attackAngle = 8f;
     [SerializeField] private float fireInterval = 0.5f;
-    [SerializeField] private float damage = 10f;
+    [SerializeField] private int damage = 10;
     [SerializeField] private float attackRange = 20f;
 
     [Header("Line Of Sight")]
